refactor gossip gossiper
