+++ logs/log1756114527.md
@@ -1,0 +1,5 @@
+# Development Log
+
+- relocated gossip message record types to dedicated `Messages.cs` for cleaner separation of message contracts
+- restructured `Gossiper` into partial classes isolating state queries, consensus/heartbeat checks, and dissemination routines while keeping constructor and options centralized
+- introduced new unit tests covering state queries, consensus evaluation, and state dissemination under `Proto.Cluster.Gossip.Tests`
